+++ src/EventListener/Dca/PageContainer.php
@@ -1,0 +1,59 @@
+<?php
+/**
+ * Contao Open Source CMS
+ *
+ * Copyright (c) 2020 Heimrich & Hannot GmbH
+ *
+ * @author  Thomas Körner <t.koerner@heimrich-hannot.de>
+ * @license http://www.gnu.org/licences/lgpl-3.0.html LGPL
+ */
+
+
+namespace HeimrichHannot\VideoBundle\EventListener\Dca;
+
+
+use Contao\DataContainer;
+use HeimrichHannot\UtilsBundle\Choice\ModelInstanceChoice;
+use HeimrichHannot\VideoBundle\Collection\VideoProviderCollection;
+
+class PageContainer
+{
+    /**
+     * @var VideoProviderCollection
+     */
+    private $videoProviderCollection;
+    /**
+     * @var ModelInstanceChoice
+     */
+    private $modelInstanceChoice;
+
+
+    /**
+     * PageContainer constructor.
+     */
+    public function __construct(VideoProviderCollection $videoProviderCollection, ModelInstanceChoice $modelInstanceChoice)
+    {
+        $this->videoProviderCollection = $videoProviderCollection;
+        $this->modelInstanceChoice = $modelInstanceChoice;
+    }
+
+    /**
+     * @param DataContainer $dc
+     * @return array
+     */
+    public function onMceVideoProviderOptionsCallback($dc)
+    {
+        return $this->videoProviderCollection->getVideoProvider();
+    }
+
+    /**
+     * @param DataContainer $dc
+     * @return mixed
+     */
+    public function onMceLocalStorageAttribute($dc)
+    {
+        return $this->modelInstanceChoice->getCachedChoices([
+            'dataContainer' => 'tl_tracking_object'
+        ]);
+    }
+}