+++ CHANGELOG.md
@@ -1,11 +1,19 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.3.0] - 2020-06-16
+- made BeforeRenderPlayer event usable
+- added AfterRenderPlayer event
+- moved privacy center integration into event subscriber
+- add privacy center fields only if privacy center is installed
+- fixed autowire exception when privacy center is not installed
+- removed multicolumneditor dependency as it is only needed when using together with privacy center
+
 ## [0.2.2] - 2020-06-16
 - modified privacy-center integration class
 
 ## [0.2.1] - 2020-06-16
--  fixed fullsize template fallback not working when no template set in root page
+- fixed fullsize template fallback not working when no template set in root page
 
 ## [0.2.0] - 2020-06-16
 - added privacy-center intergration